Add additonal fields to Audit case notes. Refactor Audit client to use the same data for audit and search.
